refactor: Rename SemanticVersionUtils -> SemanticVersion and move to .Models namespace
